requirements.txt
bidict==0.21.4
blinker==1.4
blis==0.7.5
boto3==1.20.26
botocore==1.23.26
bs4==0.0.1
CacheControl==0.12.6
cachelib==0.1.1
cachetools==4.2.2
cashfree-sdk==0.1.3
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cymem==2.0.6
dateparser==1.0.0
decorator==4.4.2
dnspython==2.1.0
dominate==2.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
eventlet==0.33.0
fasteners==0.16.3
firebase-admin==5.0.1
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Session==0.3.2
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
google-api-core==1.30.0
google-api-python-client==2.8.0
google-apitools==0.5.32
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-businessmessages==1.0.4
google-cloud==0.34.0
google-cloud-core==1.7.0
google-cloud-firestore==2.1.3
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.46.3
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.11.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
langcodes==3.3.0
libcst==0.4.1
lxml==4.7.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
msal==1.12.0
msgpack==1.0.2
murmurhash==1.0.6
mypy-extensions==0.4.3
numpy==1.22.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
parsedatetime==2.6
pathy==0.6.1
pdoc3==0.10.0
pem==21.2.0
plivo==4.21.0
preshed==3.0.6
proto-plus==1.18.1
protobuf==3.17.3
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
PyJWT==2.1.0
pyspellchecker==0.6.3
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
pandas
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.15
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
twilio==7.9.1
typer==0.4.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
visitor==0.1.3
wasabi==0.9.0
Werkzeug==2.0.1
zipp==3.4.1
